Move Delete Project modal to more sharable location
