+++ client/src/components/Project/VideoConfirmDelete/VideoConfirmDelete.scss
