+++ client/src/components/Project/VideoConfirmDelete/index.js
